requirements.txt
Django==3.1.6
djangorestframework==3.12.4
django-request==1.5.6
python-decouple==3.1
psycopg2-binary==2.8.5
bcrypt==3.2.0
argon2-cffi==20.1.0
bs4==0.0.1
Pillow==2.2.2
pylint==2.7.4
django-rest-swagger==2.2.0
coreapi==2.3.3
python-coinmarketcap==0.2